(core) added security to endpoints and fixed minor bugs present while testing

+++ src/main/java/com/gichungasoftwares/services/ProjectServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -93,35 +94,59 @@ public class ProjectServiceImpl implements ProjectService{
         return projectRepository.save(project);
     }
 
+    // get projects assigned to a user
     @Override
     public List<Project> assignedUserProjects(Long userId, ProjectStatus projectStatus) {
         List<Project> userProjects = projectRepository.findByAssignedUserId(userId);
         // user projects by status
+        System.out.println("project status -------- " + projectStatus);
         List<Project> filteredProjects = userProjects.stream().filter(
                 project -> projectStatus == null || project.getStatus().name().equalsIgnoreCase(projectStatus.toString())
         ).collect(Collectors.toList());
 
-        return null;
+        return filteredProjects;
     }
 
+    // mark project as complete
     @Override
-    public Project markProjectAsComplete(Long projectId) throws Exception {
+    public Project markProjectAsComplete(Long projectId, Long requesterId, String requesterRole) throws Exception {
         // check if project exists
         Project project = projectRepository.findById(projectId).orElseThrow(() -> new RuntimeException("Error: Project is not found"));
         if( project == null ){
             throw new Exception("project does not exist");
         }
+
+        // check for project status
+        ProjectStatus projectStatus = project.getStatus();
+        System.out.println("Project status: " + projectStatus);
+        if(projectStatus == ProjectStatus.DONE){
+            throw new Exception("Project had already been marked as complete ");
+        }
+
+        // check for admin role or project owner
+        if(!requesterId.equals(project.getAssignedUserId()) || !requesterRole.equals(("ROLE_ADMIN"))){
+            System.out.println("requesterId------ " +requesterId);
+            System.out.println("assigned user id ------ " +project.getAssignedUserId());
+            throw new RuntimeException("Error: You do not have authority to mark this project as complete or project has not been assigned!");
+        }
+
         // set project status as complete
         project.setStatus(ProjectStatus.DONE);
         return projectRepository.save(project);
     }
 
+    // delete project
     @Override
-    public void deleteProject(Long id) throws Exception {
+    public void deleteProject(Long id, String requesterRole) throws Exception {
         // check if project exists
         Project existingProject = projectRepository.findById(id).orElseThrow(() -> new RuntimeException("Error: Project not found with that id! " +id));
         if (existingProject == null) {
             throw new Exception("Project not found!");
+        }
+        // only admins can delete projects
+        if( !requesterRole.equals(("ROLE_ADMIN")) ){
+            System.out.println("requester role ------ " + requesterRole);
+            throw new RuntimeException("Error: You do not have authority to delete this project !");
         }
         projectRepository.deleteById(id);
     }
